requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
mysql-connector-python==8.0.31
